+++ shared/src/main/java/dev/tapetum/shaders/pipeline/backend/gl/ShadowTargets.java
@@ -1,6 +1,6 @@
 package dev.tapetum.shaders.pipeline.backend.gl;
 
-import com.mojang.blaze3d.opengl.GlStateManager;
+import dev.tapetum.shaders.compat.GlStateManager;
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL12;
 import org.lwjgl.opengl.GL14;